requirements.txt
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
ffmpeg==1.4
idna==2.10
imageio==2.8.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.19.0
Pillow==7.2.0
proglog==0.1.9
requests==2.24.0
tqdm==4.47.0
urllib3==1.25.9
